# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marketplace AI API",
    "version": "1.0.0",
    "description": "API documentation for the Marketplace AI platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for webhook authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "description": "User role"
          },
          "phone": {
            "type": "string",
            "description": "User's phone number"
          },
          "address": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "state": {
                "type": "string"
              },
              "country": {
                "type": "string"
              },
              "zipCode": {
                "type": "string"
              }
            }
          },
          "notificationPreferences": {
            "type": "object",
            "properties": {
              "email": {
                "type": "boolean"
              },
              "push": {
                "type": "boolean"
              },
              "sms": {
                "type": "boolean"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Listing": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Listing ID"
          },
          "title": {
            "type": "string",
            "description": "Listing title"
          },
          "description": {
            "type": "string",
            "description": "Listing description"
          },
          "price": {
            "type": "number",
            "description": "Listing price"
          },
          "type": {
            "type": "string",
            "enum": [
              "sale",
              "auction",
              "rental"
            ],
            "description": "Listing type"
          },
          "category": {
            "type": "string",
            "description": "Listing category"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of image URLs"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending",
              "sold",
              "cancelled"
            ],
            "description": "Listing status"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "location": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Point"
                ]
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "number"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Error status"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "description": "Validation errors"
          }
        }
      },
...
      "Auction": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auction ID"
          },
          "listing": {
            "$ref": "#/components/schemas/Listing"
          },
          "startPrice": {
            "type": "number",
            "description": "Starting price in cents/smallest currency unit"
          },
          "currentPrice": {
            "type": "number",
            "description": "Current highest bid amount"
          },
          "minimumBidIncrement": {
            "type": "number",
            "description": "Minimum amount for each bid increment"
          },
          "buyNowPrice": {
            "type": "number",
            "description": "Optional buy now price"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "Auction start time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "description": "Auction end time"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "ended",
              "cancelled"
            ],
            "description": "Auction status"
          },
          "bids": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "bidder": {
                  "$ref": "#/components/schemas/User"
                },
                "amount": {
                  "type": "number"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "description": "Array of bids placed"
          },
          "winner": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "bidAmount": {
                "type": "number"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            },
            "description": "Winning bid information"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BidRequest": {
        "type": "object",
        "required": [
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "description": "Bid amount in cents/smallest currency unit"
          }
        }
      },
...
      "CancelRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for cancellation"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auctions/{listingId}/bid": {
      "post": {
        "summary": "Place a bid on an auction",
        "tags": [
          "Auctions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BidRequest"
              },
              "examples": {
                "PlaceBid": {
                  "$ref": "#/components/examples/RequestExamples/PlaceBid"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bid placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "BidPlaced": {
                    "$ref": "#/components/examples/ResponseExamples/BidPlaced"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid bid or auction state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Auction or listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auctions/{listingId}/buy-now": {
      "post": {
        "summary": "Buy auction item immediately at buy now price",
        "tags": [
          "Auctions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Item purchased at buy now price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Item purchased at buy now price",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid buy now action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Auction or listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auctions/{listingId}/start": {
      "post": {
        "summary": "Start an auction",
        "tags": [
          "Auctions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Auction started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Auction started",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid auction state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Auction or listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auctions/{listingId}/end": {
      "post": {
        "summary": "End an auction",
        "tags": [
          "Auctions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Auction ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Auction ended",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid auction state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Auction or listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auctions/{listingId}/cancel": {
      "post": {
        "summary": "Cancel an auction",
        "tags": [
          "Auctions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              },
              "examples": {
                "CancelRequest": {
                  "value": {
                    "reason": "Seller withdrew the item."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auction cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Auction cancelled",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid auction state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Auction or listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auctions/{listingId}/status": {
      "get": {
        "summary": "Get auction status",
        "tags": [
          "Auctions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Auction status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auction"
                },
                "examples": {
                  "AuctionExample": {
                    "$ref": "#/components/examples/AuctionExample"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Auction or listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}